requirements.txt
python-dotenv
imaplib
email
dnspython
dkimpy
scikit-learn
pandas
streamlit
